Fix serialization of POST/PUT request body
Parameters of POST and PUT API requests were not included in request
body because of direct string comparison (method == "POST") and missing
"PUT" case in that check.

For example, I could not use the snippets API to update or create a
snippet because I was receiving 400 Bad Request (empty JSON body).