Add preheader and amp_html to template version
New optional fields are supported allowing the client to get and set the
preheader and amp_html fields on a TemplateVersionDetails object.

The amp_html field is the AMP-enabled version of an email; in cases
where AMP is not supported the email will fall back to either the HTML
or text version of the template.